crewai-orchestrator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/misjustice_crews"]

[project]
name = "misjustice-crews"
version = "0.1.0"
dependencies = [
    "crewai>=0.86.0",
    "crewai-tools>=0.25.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pytest>=8.0",
    "pyyaml>=6.0",
    "pytest-asyncio>=0.23",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]
